+++ react/NewtailMediaShelf.tsx
@@ -1,13 +1,28 @@
 /* eslint-disable no-console */
 import type { PropsWithChildren } from 'react'
-import React from 'react'
+import React, { useMemo } from 'react'
 import { defineMessages } from 'react-intl'
+import { useRuntime } from 'vtex.render-runtime'
 
 import Shelf from './components/Shelf'
 import { NewtailMediaProvider } from './hooks/useNewtailMedia'
 
 function NewtailMediaShelf(props: PropsWithChildren<NewtailMediaShelfProps>) {
-  // console.log('🔵 🟡 🔵 🟡 Carregou Newtail Media :: Shelf 🔵 🟡 🔵 🟡')
+  const { query: queryRaw } = useRuntime()
+
+  const debug = useMemo(() => queryRaw?.debug ?? null, [queryRaw])
+
+  if (debug === 'newtail') {
+    console.log(
+      '%c 🚧 🚧 🚧 🚧  NewtailMedia :: Shelf 🚧 🚧 🚧 🚧',
+      'color:#ffb450;background:#3c3584;'
+    )
+    console.log(
+      '%c 🚀 ~ NewtailMediaShelf ~ props:',
+      'color:white;background:#3c3584;',
+      props
+    )
+  }
 
   return (
     <NewtailMediaProvider adType="product" {...props}>