+++ react/settings/index.tsx
@@ -1,7 +1,13 @@
 export const tagClassnameDefault = 'newtail-sponsored-tag'
+
+// Default
 export const searchSelectorDefault =
   '.vtex-search-result-3-x-searchResultContainer #gallery-layout-container'
 
-// Same themes 
+// Locatel
 // export const searchSelectorDefault =
 //   '.vtex-search-result-3-x-searchResultContainer .vtex-search-result-3-x-gallery--gallery-main-search'
+
+// Zona sul
+// export const searchSelectorDefault =
+//   '.vtex-search-result-3-x-searchResultContainer .vtex-search-result-3-x-gallery'